refactor(splits): rename burn auxiliary to purge
