+++ x/splits/auxiliaries/purge/auxiliary.go
@@ -1,7 +1,7 @@
 // Copyright [2021] - [2022], AssetMantle Pte. Ltd. and the code contributors
 // SPDX-License-Identifier: Apache-2.0
 
-package burn
+package purge
 
 import (
 	baseHelpers "github.com/AssetMantle/modules/helpers/base"
